+++ src/2-two/lazy/plugin.js
@@ -1,0 +1,7 @@
+import lazy from './lazyParse.js'
+
+export default {
+  lib: {
+    lazy
+  }
+}